Rename files to reflect class naming and relink as needed
