+++ src/index.html
@@ -28,7 +28,7 @@
     <script type="text/javascript" src="../data/activity.js"></script>
     <script type="text/javascript" src="./User.js"></script>
     <script type="text/javascript" src="./UserRepo.js"></script>
-    <script type="text/javascript" src="./Hydration.js"></script>
+    <script type="text/javascript" src="./HydrationRepo.js"></script>
     <script type="text/javascript" src="./scripts.js"></script>
   </body>
 </html>
