+++ test/HydrationRepo-test.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const expect = chai.expect;
 
-const HydrationRepo = require('../src/Hydration');
+const HydrationRepo = require('../src/HydrationRepo');
 
 describe('HydrationRepo', () => {
   let hydrationRepo;
